Fix line wrapping with readline and coloured prompts
Readline needs to know the width of the prompt in order to wrap lines
correctly. To get this right, it has to know which character sequences
don't take up any screen space. The \001 and \002 characters (ASCII
SOH and STX) indicate the beginning and end of a nonprinting character
sequence to readline.

Fix the line wrapping by surrounding any colour escapes with these
characters.

For this to work properly, we need to do this to the prompt passed to
raw_input(), but not to any other prompts we print, otherwise the \001
and \002 characters end up going to the terminal. Refactor the call to
raw_input into a new read_line() method to achieve this.