Avoid returning Unicode strings from __repr__()
http://bugs.python.org/issue5876 says ".__repr__() is not really
allowed to return Unicode objects in Python 2.x." If it does, the
interpreter tries to encode the Unicode string according to
sys.getdefaultencoding(), which is usually ASCII.

This means that any Unicode characters in expressions might cause
UnicodeEncodeError to be raised in strange places (e.g. when trying to
debug with pdb). Avoid this by escaping Unicode characters in
__repr__().