Avoid hash-seed-dependent results in Eigenvalues tests
Eigenvalues calls into SymPy, which returns the eigenvalues as dict
keys. We sort them based on their absolute value, which might be
symbolic. In that case the sort order can depend on the original
ordering of the dict keys.

Since the ordering of dict keys depends on the hash seed, the order of
elements in the list returned by Eigenvalues can depend on the hash
seed (or the Python implementation).

Fix test failures under Travis by explicitly sorting the result of
Eigenvalues in the tests.