Stop using generator expressions to add System` to sets of symbols
Fixes TeXForm[a b] generating a Unicode invisible times character,
which TeX doesn't understand.

This problem was caused by a bug in the code to format expressions.
The check on whether a symbol was a form name was incorrectly
returning False, because Unicode string comparisons in Python work
differently in tuples and in generator expressions:

  >>> t1 = ('a', 'b')
  >>> t2 = (x for x in ('a', 'b'))
  >>> 'a' in t1
  True
  >>> u'a' in t1
  True
  >>> 'a' in t2
  True
  >>> u'a' in t2
  False

Fix this by defining functions to safely prefix sets of symbols with
System`.