Remove deprecation warning associated with \
In older Python's backslash (\) followed by a symbol that is not an by
oen of the acceptable escape characters was okay and would be the two
symbols. For example `"\["` would represent the string `"\["`.

In more recent Python, a deprecation warning is given for backslash that
isn't followed by an accepted escape character now gives a deprecation
warning. What is expected in these cases is the escape the escape, .e.g
`\\[`.