Fix runtime bugs seen from GS3.m in per_expr()
per_expr() needs to return a in interator return, not the default implied
`return None` as can happen.

In Python 3, // is used for truncated integer division. In Python 2 it
was /. That now returns float in Python 3.

Note the brokenness of self.get_wrappings

To reproduce brokenness before build, in `mathics` run:
```
   <<"https://raw.githubusercontent.com/rebcabin/Mathics/master/mathics/packages/GS3.m"
```