Set to use SansSerif by default
Having used both for a while, as others in the past have observed
SansSerif looks nicer than without.

Now that we have online docs this is all the more apparent.

(Not sure why we center, things either way, but I guess that's
different)

In discussing this with Gark, the feeling is that it would be nice if we
had a user profile for preferences like that could set.

However we don't have that right now. Given this my take is SansSerif is
a better default.