Document "hash input" button & Other doc revisions
Add hash button and move to persistance section since this is more
closely what this is about. It doesn't have anything to do with loading
or saving a file.

Some other doc changes were made as well.

Please note on formatting that line breaks are take verbatim. This is
not like say, TeX or ReStructuredText where these are ignored, but more
like Markdown where linebreaks are significant.