Remove "Protect" Attribute hackiness on Builtin
There are a number of built-ins that don't want to be Protected by
default, and we currently have no way to indicate that.

Currently "MakeBoxes" is hacked in to change this and "assignment"
has another hack to change this as well. And they do this using the
`List#clear()` function which removes *all* attributes, not just Protect.

In this change, setting the `attribute = ("Unprotect", ...)`  indicates
to Builtin we don't want to add the "Protect" attribute.

Note however this isn't a true attribute like "Protect" is - it only
means something in the declaration of a Builtin, and doesn't exist at
run time.

The need for this change becomes all the more important as we move
things into PyMathics because a growing number of these for
variable-like definition will be desired.

Another Note: in the future we may want to add a class specificaly for
variable-like definition.