Start combinatorica examples
Basically try the examples in the first page of the book.
The test harness has been move from GS0.m to test_driver.m
so we can reuse it. (It really doesn't have anything to do with
Gries-Schneider code.)

Note: GS tests were broken and to some extent still are.

Exit[1] no exits, but Quit[] does.

Also mathics -e doesn't catch Aborted and things like that and it
should.

If we caught that, in we'd see that reality GS3.m fails.