Add from_python option and bug fix
Add Mathics option OmitSurroundingQuotes and omit_surrounding_quotes
in from_python to indicate whether we want to add surrounding quotes
when converting from python.

Also we had a bug in formatting from python string to Mathics String.
Previously we slapped a double quote (") around the outside. However
that assumes internally there isn't a double quote.

Instead we now use Python's repr() which handles this

However it can produce a string that starts out with a single quote (')
but we shouldn't explect the output of PythonForm to be something that
can be then input into Mathics.