Bug in handling Sum[] min/max limits ...
Functions like Quotient[4, 2] (== 4) were not getting evaluated but
instead passed to `to_sympy()` which can't handle this as this
is purley a Python-based evaluation.

This uncovered a major weakness, that to_sympy (and probably to_python)
wasn't getting passed an evaluation so these functions, _could not_
do a proper conversion when there was an expression like this.

Bug found in Combinatorica V0.9. We can now handle up to and including section 1.4.1.

Also we blackend lists.py