Try simplying SympyFunction builtins
Would have expected to catch/DRY more items here. But I am optimistic
that by starting this as we add more Sympy functions we will be making
more use of this.

I think it is better to have one way that is used regularly than open-code this
a number of places.