Handle Compress/Uncompress checks better..
That is, in a way that doesn't depend on OS characteristics or
specifics of how Compress compresses things.

Simplify test/helper.py and remove unused imports.

BTW: I think ToString should be the exception not the rule.
In other words the defaults to_string_expr and to_string_expected should
be opposite of what it currently is.