Introduce Integer1 = Integer(1)
@mmatera: a while back in branch not pushed, I also had other numbers
and in particular Integer(0) which is even more common.

I had problems somewhere along the line with that, so I didn't commit
any of this.

Since you introduced more locally IntegerOne, put this in for that one
case.

However rather than do more Integer symbols in this branch, let's merge
it in and do this as a follow on.

The number of changes in this branch is getting quite long.