Pull out more from graphics!
Refactor all of the color stuff into drawing/colors.py
and color_internals.py

Note drawing/color_internals.py used to be somewhat vague name
drawing/colors.py. But really it just had internal color routines.

drawing/colors.py was selected because that's what the Section name is
in WL https://reference.wolfram.com/language/guide/Colors.html and the
classes in there all into what's described there.

Down the line it would be nice to separate better externally visable
classes from all the implementaiton code that is not visiable, but
that's too big a lift for now.

Note: this started out as just an effort to document some of the
builtins that had no documentaiton.

And a little of that has been done too.