Reinstate Density plot for SVG - rough cut
Back in a less-hacky way and where we *describe* what's up.

Looks are a little ugly but can probably be fixed using rectangular gradients
described in comments in density_plot_box().

Also, we now produce more human-readable SVG. Since this is produced
by hand, it is useful to facilitate humans debugging this.