Fix some of the Doc breaking in previous work
common_doc.py: Black messed up the testcase RE.
Rename generic Doc to XMLDoc since that seems to be more what it's about
doc2latex.py: be more verbose about what it's doing with an option to be
silent.

There is still a bit more work to do here. I think generally we are
closer though.