Track regexp test change...
Our regular expression match for tests now requires a space at the
beginning of the line.

The regular expression change was done to assist matching tests and
having them be distinct from non-text documentation.

Alas this means that are static manual examples which didn't have the
leading space must change.