Revise doc data for TeX
We now keep both a version of the internal docs for LaTeX formatting
in both package space and user space.

User space is writable and can be updated with additional modules as
might be found in PyMathics modules one day

The system space is a fallback and is supplied by the distribution.
It has document data at the time the distribution was packaged.

Note that data destined for HTML needs to be separate.

The main difference is that the output results format are different:
LaTeX/asymptote versus HTML/MathJax/ThreeJS.

Simple, huh?